kvstore: fix parameter propagation

+++ GimbalFirmware/Headers/parameters/mavlink_parameter_interface.h
@@ -90,7 +90,7 @@ typedef struct {
 void init_default_mavlink_params();
 void handle_param_set(mavlink_message_t* received_msg, MotorDriveParms* md_parms);
 void handle_param_read(mavlink_message_t* received_msg);
-void send_gimbal_param(int param_num);
+void send_gimbal_param(uint16_t param_num);
 void gimbal_param_update(GimbalMavlinkParameterID param_id, Uint32 value, ControlBoardParms *cb_parms);
 
 extern float send_torques;
